создан модуль для рендера задач
